requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
DateTime==5.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
git-filter-repo==2.38.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.25.2
packaging==23.1
pdfkit==1.0.0
Pillow==10.0.0
protobuf==4.21.12
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
wkhtmltopdf==0.2
WTForms==3.0.1
zope.interface==6.0
